requirements.txt
pandas
openpyxl
azure-storage-file-datalake
snowflake-connector-python
python-dotenv